+++ test/integration/assign_test.rb
