+++ test/integration/blank_test.rb
