+++ test/integration/capture_test.rb
