+++ test/integration/drop_test.rb
