+++ test/integration/error_handling_test.rb
