+++ test/integration/hash_ordering_test.rb
