+++ test/integration/output_test.rb
