+++ test/integration/standard_filter_test.rb
