+++ test/integration/tags/continue_tag_test.rb
