+++ test/integration/tags/include_tag_test.rb
