+++ test/integration/tags/increment_tag_test.rb
