+++ test/integration/tags/raw_tag_test.rb
