+++ test/unit/tags/case_tag_unit_test.rb
@@ -1,10 +1,10 @@
 require 'test_helper'
 
-class CaseTagTest < Test::Unit::TestCase
+class CaseTagUnitTest < Test::Unit::TestCase
   include Liquid
 
   def test_case_nodelist
     template = Liquid::Template.parse('{% case var %}{% when true %}WHEN{% else %}ELSE{% endcase %}')
     assert_equal ['WHEN', 'ELSE'], template.root.nodelist[0].nodelist
   end
-end # CaseTest
+end
