+++ test/unit/tags/if_tag_unit_test.rb
@@ -1,0 +1,8 @@
+require 'test_helper'
+
+class IfTagUnitTest < Test::Unit::TestCase
+  def test_if_nodelist
+    template = Liquid::Template.parse('{% if true %}IF{% else %}ELSE{% endif %}')
+    assert_equal ['IF', 'ELSE'], template.root.nodelist[0].nodelist
+  end
+end
